Added requests to setup.py
